changed the datatype to double

+++ .ignore/src/main/resources/Rectangle/src/main/java/com/thoughtworks/rectangle/Rectangle.java
@@ -1,14 +1,14 @@
 package com.thoughtworks.rectangle;
 public class Rectangle{
-    private final int length;
-    private final int breadth;
+    private final double length;
+    private final double breadth;
 
-    public Rectangle(int length,int breadth){
+    public Rectangle(double length,double breadth){
         this.length=length;
         this.breadth=breadth;
     }
 
-    public int area(){
+    public double area(){
         return length*breadth;
     }
 }
